requirements.txt
streamlit>=1.32.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
chardet>=5.2.0
transformers>=4.38.0
torch>=2.2.0
tokenizers>=0.15.0
sentence-transformers>=2.5.0
scikit-learn>=1.4.0
joblib>=1.3.0
chromadb>=0.4.22
nltk>=3.8.1
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0
plotly>=5.19.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0
